added segregation
