+++ server/.gitignore
@@ -5,4 +5,4 @@
 .env 
 
 
-/node_modules
+node_modules
