+++ server/models/FriendRequest.js
@@ -1,10 +1,11 @@
-const mongoose = require("mongoose");
 
-const FriendRequestSchema = new mongoose.Schema({
-  senderId: { type: mongoose.Schema.Types.ObjectId, ref: "User", required: true },
-  receiverId: { type: mongoose.Schema.Types.ObjectId, ref: "User", required: true },
-  status: { type: String, enum: ["pending", "accepted", "denied"], default: "pending" },
-  createdAt: { type: Date, default: Date.now },
+const mongoose = require('mongoose');
+
+const friendRequestSchema = new mongoose.Schema({
+  sender_id: { type: mongoose.Schema.Types.ObjectId, ref: 'User', required: true, index: true },
+  receiver_id: { type: mongoose.Schema.Types.ObjectId, ref: 'User', required: true, index: true },
+  status: { type: String, enum: ['pending', 'accepted', 'declined'], default: 'pending' },
+  created_at: { type: Date, default: Date.now },
 });
 
-module.exports = mongoose.model("FriendRequest", FriendRequestSchema);
+module.exports = mongoose.model('FriendRequest', friendRequestSchema);
