+++ server/models/User.js
@@ -1,15 +1,9 @@
-const mongoose = require("mongoose");
-const bcrypt = require("bcryptjs");
+const mongoose = require('mongoose');
 
-const UserSchema = new mongoose.Schema({
-  username: { type: String, required: true, unique: true },
+const userSchema = new mongoose.Schema({
+  username: { type: String, unique: true, required: true, index: true },
   password: { type: String, required: true },
+  created_at: { type: Date, default: Date.now },
 });
 
-UserSchema.pre("save", async function (next) {
-  if (!this.isModified("password")) return next();
-  this.password = await bcrypt.hash(this.password, 10);
-  next();
-});
-
-module.exports = mongoose.model("User", UserSchema);
+module.exports = mongoose.model('User', userSchema);
