+++ server/package.json
@@ -4,7 +4,7 @@
   "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon index"
+    "start": "nodemon server.js"
   },
   "keywords": [],
   "author": "",
@@ -17,8 +17,8 @@
     "express": "^4.21.2",
     "http": "^0.0.1-security",
     "jsonwebtoken": "^9.0.2",
-    "mongodb": "^6.13.0",
-    "mongoose": "^8.13.0",
+    "mongodb": "^6.15.0",
+    "mongoose": "^8.13.1",
     "mysql2": "^3.11.5",
     "nodemon": "^3.1.7",
     "socket.io": "^4.8.1"
