+++ server/routes/users.js
@@ -1,0 +1,17 @@
+const express = require('express');
+const User = require('../models/User');
+const authenticateToken = require('../middleware/auth');
+
+const router = express.Router();
+
+// Fetch All Users
+router.get('/', authenticateToken, async (req, res) => {
+  try {
+    const users = await User.find({ _id: { $ne: req.userId } }, 'username');
+    res.json(users);
+  } catch (err) {
+    res.status(500).json({ message: 'Error fetching users' });
+  }
+});
+
+module.exports = router;
